requirements.txt
pygame
python-dotenv
numba